+++ collectorpb/collectorpbfakes/fake_collector_service_client.go
@@ -4,8 +4,9 @@ package collectorpbfakes
 import (
 	"sync"
 
-	"github.com/lightstep/lightstep-tracer-go/collectorpb"
+	collectorpb "github.com/lightstep/lightstep-tracer-protos/go/lightstep/collector"
 	"golang.org/x/net/context"
+	_ "google.golang.org/genproto/googleapis/api/annotations"
 	"google.golang.org/grpc"
 )